+++ kod/src/main/java/com/koreait/kod/model/productAndWishlist/ProductDAO.java
@@ -1,11 +1,11 @@
-package com.koreait.kod.model.wishlist;
+package com.koreait.kod.model.productAndWishlist;
 
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 
-public class WishlistDAO {
+public class ProductDAO {
 	@Autowired // DI의존주입 /* @Autowired는 만능이 아니다.. 메모리에 로드가 되어있어야 가능하다. -> new JdbcTemplate();
 	private JdbcTemplate jdbcTemplate; // 의존관계 -> DI(의존주입) -> @Autowired
 
@@ -15,23 +15,23 @@ public class WishlistDAO {
 	private static final String UPDATE="";
 	private static final String DELETE="";
 
-	public List<WishlistDTO> selectAll(WishlistDTO wishlistDTO) {
+	public List<ProductDTO> selectAll(ProductDTO productDTO) {
 		return null;
 	}
 
-	public WishlistDTO selectOne(WishlistDTO wishlistDTO) {
+	public ProductDTO selectOne(ProductDTO productDTO) {
 		return null;
 	}
 
-	public boolean insert(WishlistDTO wishlistDTO) {
+	public boolean insert(ProductDTO productDTO) {
 		return false;
 	}
 
-	public boolean update(WishlistDTO wishlistDTO) {
+	public boolean update(ProductDTO productDTO) {
 		return false;
 	}
 
-	public boolean delete(WishlistDTO wishlistDTO) {
+	public boolean delete(ProductDTO productDTO) {
 		return false;
 	}
 }
