+++ kod/src/main/java/com/koreait/kod/model/productAndWishlist/WishListDAO.java
@@ -1,11 +1,11 @@
-package com.koreait.kod.model.product;
+package com.koreait.kod.model.productAndWishlist;
 
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 
-public class ProductDAO {
+public class WishListDAO {
 	@Autowired // DI의존주입 /* @Autowired는 만능이 아니다.. 메모리에 로드가 되어있어야 가능하다. -> new JdbcTemplate();
 	private JdbcTemplate jdbcTemplate; // 의존관계 -> DI(의존주입) -> @Autowired
 
@@ -15,23 +15,23 @@ public class ProductDAO {
 	private static final String UPDATE="";
 	private static final String DELETE="";
 
-	public List<ProductDTO> selectAll(ProductDTO productDTO) {
+	public List<WishListDTO> selectAll(WishListDTO wishListDTO) {
 		return null;
 	}
 
-	public ProductDTO selectOne(ProductDTO productDTO) {
+	public WishListDTO selectOne(WishListDTO wishListDTO) {
 		return null;
 	}
 
-	public boolean insert(ProductDTO productDTO) {
+	public boolean insert(WishListDTO wishListDTO) {
 		return false;
 	}
 
-	public boolean update(ProductDTO productDTO) {
+	public boolean update(WishListDTO wishListDTO) {
 		return false;
 	}
 
-	public boolean delete(ProductDTO productDTO) {
+	public boolean delete(WishListDTO wishListDTO) {
 		return false;
 	}
 }
